+++ truffle/src/com.oracle.truffle.api.test/src/com/oracle/truffle/api/test/builtin/BuiltinTestObject.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2012, 2018, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2019, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The Universal Permissive License (UPL), Version 1.0
@@ -38,63 +38,27 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package com.oracle.truffle.api.dsl;
+package com.oracle.truffle.api.test.builtin;
 
-import com.oracle.truffle.api.nodes.Node;
-import java.util.List;
+import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
+import com.oracle.truffle.api.dsl.Specialization;
 
-/**
- * Enables the dynamic creation of generated nodes. It provides an convenient way to instantiate
- * generated node classes without using reflection.
- *
- * @since 0.8 or earlier
- */
-public interface NodeFactory<T> {
+public final class BuiltinTestObject extends BuiltinObject {
 
-    /**
-     * Instantiates the node using the arguments array. The arguments length and types must suffice
-     * one of the returned signatures in {@link #getNodeSignatures()}. If the arguments array does
-     * not suffice one of the node signatures an {@link IllegalArgumentException} is thrown.
-     *
-     * @param arguments the argument values
-     * @return the instantiated node
-     * @throws IllegalArgumentException
-     * @since 0.8 or earlier
-     */
-    T createNode(Object... arguments);
+    private static final BuiltinDescriptor DESCRIPTOR = describe(BuiltinTestObjectFactory.getFactories());
 
-    /**
-     * Returns the node class that will get created by {@link #createNode(Object...)}. The node
-     * class does not match exactly to the instantiated object but they are guaranteed to be
-     * assignable.
-     *
-     * @since 0.8 or earlier
-     */
-    Class<T> getNodeClass();
+    @Override
+    protected BuiltinDescriptor getBuiltinDescriptor() {
+        return DESCRIPTOR;
+    }
 
-    /**
-     * Returns a list of signatures that can be used to invoke {@link #createNode(Object...)}.
-     *
-     * @since 0.8 or earlier
-     */
-    List<List<Class<?>>> getNodeSignatures();
-
-    /**
-     * Returns a list of children that will be executed by the created node. This is useful for base
-     * nodes that can execute a variable amount of nodes.
-     *
-     * @since 0.8 or earlier
-     */
-    List<Class<? extends Node>> getExecutionSignature();
-
-    /**
-     * Returns the uncached version of this node or <code>null</code> if {@link GenerateUncached}
-     * was not applied to the node.
-     *
-     * @since 20.0.0 beta 2
-     */
-    default T getUncachedInstance() {
-        return null;
+    @Builtin
+    abstract static class TestMethod extends BuiltinNode {
+        @Specialization
+        @TruffleBoundary
+        Object doDefault(@SuppressWarnings("unused") BuiltinTestObject receiver, String argument) {
+            return "test" + argument;
+        }
     }
 
 }